refactor: Remove LatticeSpace and Move the functions to VoxelSpaceBase
